Task Management System/requirements.txt
flask==3.0.3
sqlalchemy==2.0.32
bcrypt==4.2.0
pyjwt==2.9.0